style: add new play-from-scratch-button component
