+++ boolebot/src/Components/Gameplay/PlayFromScratchBtn.jsx
@@ -1,0 +1,24 @@
+import React from 'react'
+import { Link } from 'react-router-dom';
+
+export default function PlayFromScratchBtn({setBotsArray}) {
+    
+  function restartGame() {
+    setBotsArr([]);
+  }
+    return (
+    <div>
+      <Link to="/">
+        <div>
+          <button
+            onClick={() => {
+              restartGame();
+            }}
+          >
+            Play From Scratch
+          </button>
+        </div>
+      </Link>
+    </div>
+  );
+}
